adjust for esp-idf-sys size_t u32->usize

+++ esp-idf-sys/src/alloc.rs
@@ -20,11 +20,11 @@ unsafe impl Sync for Esp32Alloc {}
 
 unsafe impl GlobalAlloc for Esp32Alloc {
     unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        heap_caps_malloc(layout.size() as u32, MALLOC_CAP_8BIT as _) as *mut _
+        heap_caps_malloc(layout.size(), MALLOC_CAP_8BIT as _) as *mut _
     }
 
     unsafe fn realloc(&self, ptr: *mut u8, _layout: Layout, new_size: usize) -> *mut u8 {
-        heap_caps_realloc(ptr as *mut _, new_size as u32, MALLOC_CAP_8BIT as _) as *mut _
+        heap_caps_realloc(ptr as *mut _, new_size, MALLOC_CAP_8BIT as _) as *mut _
     }
 
     unsafe fn dealloc(&self, ptr: *mut u8, _layout: Layout) {
